SWD Batching for JLink
This commit implements batching for register writes in the SWD JLink
driver. This allows to significantly increase throughput for flashing
and debug operations.

This commit currently depends on a fix for the JayLink crate. The
split_off function has a bug and applies an invalid bit offset for the
self BitIter instance.

Fixed in:
jonas-schievink/jaylink@4213b5cb2d49d38544c34da6059bc42a07083092